requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.4.0+cpu
anyio==4.4.0
faiss-cpu==1.8.0
fastapi==0.114.0
gitpython==3.1.43
httpx==0.27.0
numpy==1.26.4
openai==1.40.0
pydantic==2.8.2
pydantic-settings==2.3.4
pytest==8.3.1
pytest-asyncio==0.23.7
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.0.1
streamlit==1.37.0
uvicorn[standard]==0.30.1
